Added post show, userpostcontroller, components for posts

+++ app/Http/Controllers/PostController.php
@@ -10,9 +10,14 @@ class PostController extends Controller
     public function index()
     {
         $posts = Post::orderBy('created_at', 'desc')->with(['user', 'likes'])->paginate(20);
-        
+
         return view('posts/index', [
             'posts' => $posts
+        ]);
+    }
+    public function show(Post $post) {
+        return view('posts.show', [
+            'post'=>$post
         ]);
     }
     public function store(Request $request)
@@ -20,12 +25,13 @@ class PostController extends Controller
         $this->validate($request, [
             'body' => 'required'
         ]);
-        $request->user()->posts()->create($request->only('body'));
-
+        //$request->user()->posts()->create($request->only('body'));
+            dd($request->user()->posts());
         return back();
     }
     public function destroy(Post $post)
     {    
+        $this->authorize('delete', $post);
         $post->delete();
         return back();
     }
